+++ ingester/src/lib.rs
@@ -28,6 +28,7 @@ pub mod lifecycle;
 mod poison;
 pub mod querier_handler;
 pub(crate) mod query_adaptor;
+mod sequence_range;
 pub mod server;
 pub(crate) mod stream_handler;
 #[cfg(test)]
